requirements.txt
discord.py>=2.3
aiofiles
python-dotenv
tqdm
pytest
